move websites-server to src/server
