+++ nodemon.json
@@ -1,0 +1,4 @@
+{
+  "watch": ["src/server"],
+  "exec": "node src/server/index.js"
+}
