+++ package.json
@@ -11,7 +11,8 @@
     "doc": "LIMIT_BLOG_PAGES=1 gatsby develop",
     "author": "LIMIT_BLOG_PAGES=1 SKIP_DOCS=true dvc pull && gatsby develop",
     "build": "dvc pull --force && gatsby build",
-    "start": "node ./src/server.js",
+    "start": "node ./src/server/index.js",
+    "start-dev": "nodemon",
     "heroku-postbuild": "./scripts/heroku-deploy.sh",
     "test": "jest",
     "format": "prettier --write",
@@ -50,17 +51,20 @@
     "@aws-sdk/client-cloudfront": "3.840.0",
     "@aws-sdk/client-s3": "3.840.0",
     "@dvcorg/gatsby-theme": "workspace:^",
-    "@dvcorg/websites-server": "workspace:^",
+    "@hapi/wreck": "18.1.0",
     "@iframe-resizer/react": "5.4.6",
-    "@octokit/request": "8.4.1",
+    "@octokit/request": "9.2.4",
     "@radix-ui/react-dialog": "1.1.14",
     "@radix-ui/react-label": "2.1.7",
     "@reach/router": "1.3.4",
     "@reach/tooltip": "0.18.0",
+    "@sentry/node": "9.33.0",
     "algoliasearch": "4.25.2",
     "clsx": "2.1.1",
+    "compression": "1.8.0",
     "date-fns": "4.1.0",
     "dotenv": "16.6.1",
+    "express": "5.1.0",
     "fs-extra": "11.3.0",
     "gatsby": "5.14.5",
     "gatsby-plugin-catch-links": "5.14.0",
@@ -76,11 +80,16 @@
     "gatsby-remark-images": "7.14.0",
     "gatsby-transformer-remark-frontmatter": "1.1.0",
     "gatsby-transformer-sharp": "5.14.0",
+    "helmet": "8.1.0",
+    "http-proxy-middleware": "3.0.5",
     "isomorphic-fetch": "3.0.0",
     "lucide-react": "0.525.0",
+    "mime-types": "3.0.1",
     "mkdirp": "3.0.1",
     "moment": "2.30.1",
     "nanoid": "5.1.5",
+    "node-cache": "5.1.2",
+    "permissions-policy": "0.6.0",
     "promise-polyfill": "8.3.0",
     "prop-types": "15.8.1",
     "raf-polyfill": "1.0.0",
@@ -95,6 +104,7 @@
     "react-use": "17.6.0",
     "repo-link-check": "0.14.2",
     "s3-sync-client": "4.3.1",
+    "serve-handler": "6.1.6",
     "slick-carousel": "1.8.1",
     "tailwind-merge": "3.3.1",
     "tailwindcss-animate": "1.0.7",
@@ -105,6 +115,7 @@
   },
   "devDependencies": {
     "@babel/core": "7.27.7",
+    "@eslint/eslintrc": "3.3.1",
     "@eslint/js": "9.30.0",
     "@eslint/json": "0.12.0",
     "@types/gatsbyjs__reach-router": "2.0.5",
@@ -122,8 +133,10 @@
     "eslint": "9.30.0",
     "eslint-config-prettier": "10.1.5",
     "eslint-import-resolver-typescript": "4.4.4",
+    "eslint-plugin-import": "2.32.0",
     "eslint-plugin-import-x": "4.16.1",
     "eslint-plugin-jsx-a11y": "6.10.2",
+    "eslint-plugin-n": "17.20.0",
     "eslint-plugin-prettier": "5.5.1",
     "eslint-plugin-react": "7.37.5",
     "eslint-plugin-react-hooks": "5.2.0",
@@ -133,6 +146,7 @@
     "husky": "9.1.7",
     "jest": "29.7.0",
     "lint-staged": "16.1.2",
+    "nodemon": "3.1.10",
     "prettier": "3.6.2",
     "remark": "15.0.1",
     "remark-html": "16.0.1",
