--- packages/websites-server/server.js
@@ -1,4 +1,0 @@
-const server = require('./src/index')
-
-// run server
-server.run()
