+++ src/gatsby/models/github/index.js
@@ -1,4 +1,9 @@
-const { request } = require('@octokit/request')
+let request
+;(async () => {
+  const mod = await import('@octokit/request')
+  request = mod.request
+})()
+
 /*
    GitHub Static Data
