+++ src/server/middleware/redirects/index.js
