+++ src/server/middleware/serve/s3.js
